requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
aiofiles==23.2.1
python-dotenv==1.0.1
supabase==2.4.2
asyncpg==0.29.0
redis==5.0.4
python-dateutil==2.9.0
pytz==2024.1
validators==0.28.1
